Add Copy_constructor class with copy constructor implementation
This commit adds a new class called Copy_constructor.java which demonstrates the implementation of a copy constructor in Java. The class includes a copy constructor that creates a new object by copying the values of another object. The purpose of this commit is to provide an example of how to use a copy constructor to create a new object with the same values as an existing object.